호스트가 매칭방 수정하기 기능 추가

+++ bootProject/src/main/java/com/busanit501/bootproject/controller/PetController.java
@@ -84,15 +84,15 @@ public class PetController {
      * 반려동물 응답용 DTO
      */
     static class PetResponse {
-        private Integer id;
+        private Long id;
         private String name;
 
-        public PetResponse(Integer id, String name) {
+        public PetResponse(Long id, String name) {
             this.id = id;
             this.name = name;
         }
 
-        public Integer getId() {
+        public Long getId() {
             return id;
         }
