+++ bootProject/src/main/java/com/busanit501/bootproject/dto/PetDTO.java
@@ -18,7 +18,7 @@ public class PetDTO {
 
     @NotNull(message = "반려동물 나이는 필수 입력 항목입니다.")
     @Min(value = 0, message = "반려동물 나이는 0 이상이어야 합니다.")
-    private Integer age;
+    private Long age;
 
     @NotNull(message = "반려동물 성별은 필수 입력 항목입니다.")
     private Gender gender;
