+++ bootProject/src/main/java/com/busanit501/bootproject/repository/MatchingRoomRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatchingRoomRepository extends JpaRepository<MatchingRoom, Integer> {
+public interface MatchingRoomRepository extends JpaRepository<MatchingRoom, Long> {
     // 추가적인 쿼리 메서드가 필요하면 여기에 정의
 }
